feat: Add mailListContainer component
